move stuff around, use envs
