+++ shibbo/.env.example
@@ -1,3 +1,8 @@
-HY_ORGANIZATION_SECRET=secret
-HY_ORGANIZATION_ID=12346
 PORT=6202
+FRONTEND_URL=http://localhost:3000
+BACKEND_URL=http://localhost:4000
+API_URL=http://localhost:4000/api
+AUTH_URL=http://localhost:4000/auth
+
+HY_IDP_URL=http://localhost:7000/saml/sso
+HAKA_IDP_URL=http://localhost:7002/saml/sso
