+++ shibbo/src/handlers/index.ts
@@ -1,3 +1,13 @@
-export * from "./connect"
-export * from "./signIn"
-export * from "./signUp"
+import { HandlerCallback } from "./callback"
+import { connectHandler } from "./connect"
+import { signInHandler } from "./signIn"
+import { signUpHandler } from "./signUp"
+
+export const handlers: Record<string, HandlerCallback> = {
+  "sign-in": signInHandler,
+  "sign-up": signUpHandler,
+  connect: connectHandler,
+}
+
+export * from "./callback"
+export { connectHandler, signInHandler, signUpHandler }
