+++ shibbo/src/server.ts
@@ -1,38 +1,28 @@
 import cors from "cors"
 import express, {
   CookieOptions,
-  RequestHandler,
   urlencoded,
 } from "express"
-import * as fs from "fs"
 import morgan from "morgan"
-import passport, { AuthenticateOptions } from "passport"
+import passport from "passport"
 import { MultiSamlStrategy } from "passport-saml"
 import shibbolethCharsetMiddleware from "unfuck-utf8-headers-middleware"
 
 import {
-  BACKEND_URL,
   DOMAIN,
-  FRONTEND_URL,
+  HAKA_IDP_URL,
+  HY_IDP_URL,
+  MOOCFI_CERTIFICATE,
   PORT,
   SHIBBOLETH_HEADERS,
 } from "./config"
+import { callbackHandler } from "./handlers"
 import {
-  connectHandler,
-  signInHandler,
-  signUpHandler,
-} from "./handlers"
-import { HandlerCallback } from "./types/handlers"
-import {
-  decodeRelayState,
+  convertObjectKeysToLowerCase,
   encodeRelayState,
-} from "./util/relayState"
+} from "./util"
 
 const isProduction = process.env.NODE_ENV === "production"
-
-if (isProduction && (!BACKEND_URL || !FRONTEND_URL)) {
-  throw new Error("BACKEND_URL and FRONTEND_URL must be set")
-}
 
 const app = express()
 
@@ -40,7 +30,7 @@ app.set("port", PORT)
 app.use(cors())
 app.use(express.json())
 app.use((req, _, next) => {
-  if (!isProduction && !req.headers.schacpersonaluniquecode) {
+  if (!isProduction && !req.headers.edupersonprincipalname) {
     req.headers = { ...req.headers /*, ...defaultHeaders */ }
   }
 
@@ -72,70 +62,10 @@ app.use((req, res, next) => {
 })
 app.use(passport.initialize())
 
-/*const connectHandler: RequestHandler = (req, res, next) => {
-  const relayState = encodeRelayState(req)
-
-  passport.authenticate(
-    "multi",
-    {
-      additionalParams: {
-        RelayState: relayState
-      },
-    } as AuthenticateOptions,
-    (err, user) => {
-      console.log("connect handler", err, user, req.query)
-
-      req.login(user, (err) => {
-        if (err) {
-          console.log("connect error", err)
-        }
-      })
-
-      const { provider, action } = req.params
-      const { language = "en" } = req.query
-
-      res.redirect(`/foo?${Object.entries({ provider, action, language }).map(([key, value]) => `${key}=${value}`).join("&")}`)
-    },
-  )(req, res, next)
-}*/
-
-const handlers: Record<string, HandlerCallback> = {
-  "sign-in": signInHandler,
-  "sign-up": signUpHandler,
-  connect: connectHandler,
-}
 
 const providers: Record<string, string> = {
-  hy: "http://localhost:7000/saml/sso",
-  haka: "http://localhost:7002/saml/sso",
-}
-
-const callbackHandler: RequestHandler = (req, res, next) => {
-  const relayState = req.params.RelayState || req.body.RelayState
-  const action = req.params.action || decodeRelayState(relayState)?.action
-
-  const decodedRelayState = decodeRelayState(
-    req.params.RelayState || req.body.RelayState,
-  ) ?? {
-    action: req.params.action,
-    provider: req.params.provider,
-    language: ((req.query.language || req.params.language) as string) ?? "en",
-  }
-  console.log("callback relaystate", decodedRelayState)
-
-  if (!Object.keys(handlers).includes(action)) {
-    throw new Error(`unknown action ${action}`) // TODO: something more sensible
-  }
-
-  passport.authenticate(
-    "hy-haka",
-    {
-      additionalParams: {
-        RelayState: relayState
-      }
-    } as AuthenticateOptions,
-    handlers[action](req, res, next)
-  )(req, res, next)
+  hy: HY_IDP_URL,
+  haka: HAKA_IDP_URL,
 }
 
 /*passport.serializeUser((user, done) => {
@@ -163,11 +93,7 @@ passport.use(
           path: `/callbacks/${provider}/${action}/${language}`,
           entryPoint: providers[provider],
           issuer: "https://mooc.fi/sp",
-          cert: fs
-            .readFileSync(
-              __dirname + "/../shibboleth-staging/certs/mooc.fi.crt",
-            )
-            .toString(),
+          cert: MOOCFI_CERTIFICATE,
           authnRequestBinding: "HTTP-POST",
           forceAuthn: true,
           identifierFormat:
@@ -182,10 +108,7 @@ passport.use(
       console.log("got profile", profile)
       done(
         null,
-        Object.entries(profile?.attributes).reduce(
-          (acc, [key, value]) => ({ ...acc, [key.toLowerCase()]: value }),
-          {},
-        ) ?? undefined,
+        convertObjectKeysToLowerCase(profile?.attributes)
       )
     },
   ),
